+++ src/main/java/org/beanpod/switchboard/entity/SwitchBoardUserDetails.java
@@ -3,11 +3,20 @@ package org.beanpod.switchboard.entity;
 import java.util.Collection;
 import java.util.Collections;
 import javax.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
+@Builder
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
 public class SwitchBoardUserDetails implements UserDetails {
 
   @NotNull private String username;
