+++ src/test/java/org/beanpod/switchboard/fixture/UserFixture.java
@@ -1,11 +1,15 @@
 package org.beanpod.switchboard.fixture;
 
 import org.beanpod.switchboard.dto.UserDto;
+import org.beanpod.switchboard.entity.SwitchBoardUserDetails;
 import org.beanpod.switchboard.entity.UserEntity;
 import org.openapitools.model.UserModel;
 
 public class UserFixture {
 
+  public static SwitchBoardUserDetails getUserDetails(){
+    return new SwitchBoardUserDetails().builder().username("moh@gmail.com").password("1234.").build();
+  }
   public static UserModel getUserModel() {
     return new UserModel().username("moh@gmail.com").password("1234.");
   }
